Screens refactor. Each screen has its own folder.
