+++ src/screens/index.ts
@@ -6,17 +6,17 @@ export { default as ForgotPassword }  from './auth/ForgotPassword/ForgotPassword
 export { default as ResetPassword } from './auth/ResetPassword/ResetPassword.tsx';
 
 // Screens generales
-export { default as LoadScreen } from './LoadScreen.tsx';
-export { default as WelcomeScreen } from '../screens/WelcomeScreen.tsx';
+export { default as LoadScreen } from './LoadScreen/LoadScreen.tsx';
+export { default as WelcomeScreen } from '../screens/WelcomeScreen/WelcomeScreen.tsx';
 
 // Screens del botón Inicio
 export { default as HomeScreen } from './HomeScreen/HomeScreen.tsx';
 
 // Screens del botón Grupos
-export { default as CreateGroup } from '../screens/CreateGroup.tsx';
-export { default as CreateEvent } from '../screens/CreateEvent.tsx';
-export { default as GroupInformation } from '../screens/GroupInformation.tsx';
-export { default as JoinGroup } from '../screens/JoinGroup.tsx';
+export { default as CreateGroup } from '../screens/CreateGroup/CreateGroup.tsx';
+export { default as CreateEvent } from '../screens/CreateEvent/CreateEvent.tsx';
+export { default as GroupInformation } from '../screens/GroupInformation/GroupInformation.tsx';
+export { default as JoinGroup } from '../screens/JoinGroup/JoinGroup.tsx';
 
 // Screens del botón Chatbot
 export { default as ChatbotScreen } from './ChatbotScreen/ChatbotScreen.tsx';
